requirements.txt
numpy
gymnasium
pygame
torch
matplotlib
tqdm
gym
stable-baselines3
laplace-torch
imageio